pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "aiNewReader"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.34",
    "google-genai>=0.7",
    "openai>=1.40",
    "httpx>=0.27",
    "fastfeedparser>=0.1",
    "trafilatura>=1.12",
    "rapidfuzz>=3.9",
    "lancedb>=0.10",
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "click>=8.1",
    "fastapi>=0.112",
    "uvicorn>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
]

[tool.setuptools.packages.find]
where = ["src"]
